Added a ngbCollapse directive to the element with the "navbar-collapse" CSS class.
Added Removed all of the basic unnecessary attributes in the div elements that define the navbar, in the navbar template. Implementation of the ngbCollapse directive and management of the "isMenuCollapsed" property handles the collapse of the menu-items on click during the mobile mode. (Note: default/vanilla navbar does not work in angular"

Added a click event handler to each link and set the property on the component to true to hid the menu.

Added a click event handler for the button  element with the "navbar-toggler" CSS class to toggle the value of the isMenuCollapsed property on click.